IGNITE-18130 Added possibility to override KeyManager initialization for SslContextFactory.
